+++ src/main/java/com/samatov/inventoryservicebar/services/JuiceService.java
@@ -1,6 +1,9 @@
 package com.samatov.inventoryservicebar.services;
 
-import com.samatov.inventoryservicebar.entities.Juice;
+import com.samatov.inventoryservicebar.common.exceptions.notfoundexceptions.JuiceNotFoundException;
+import com.samatov.inventoryservicebar.dto.JuiceDTO;
+import com.samatov.inventoryservicebar.entities.JuiceEntity;
+import com.samatov.inventoryservicebar.mappers.JuiceMapper;
 import com.samatov.inventoryservicebar.repositories.JuiceRepository;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
@@ -10,6 +13,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -17,48 +21,56 @@ import java.util.Optional;
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class JuiceService {
 
-
     JuiceRepository juiceRepository;
 
-    public List<Juice> getAllJuices() {
-        log.info("Всю позицию соков получили из база данных");
-        return juiceRepository.findAll();
+    JuiceMapper juiceMapper;
+
+    public List<JuiceDTO> getAllJuices() {
+        List<JuiceEntity> juices = juiceRepository.findAll();
+        log.info("Получены все позиции соков из база данных");
+        if(juices.isEmpty()) {
+            throw new JuiceNotFoundException("Позиции соков не найдены в базе данных");
+        }
+        return juices.stream().map(juiceMapper::toDto).collect(Collectors.toList());
     }
 
-    public void saveJuice(Juice juice) {
-        log.info("Сок добавлен в базу данных");
+    public void saveJuice(JuiceDTO juiceDTO) {
+        JuiceEntity juice = juiceMapper.toEntity(juiceDTO);
+        log.info("Добавление сок в базу данных");
         juiceRepository.save(juice);
     }
 
     public void deleteJuiceById(String id) {
-        log.info("Сок удален из база данных");
+        log.info("Удаление сока из базы данных");
+        if (!juiceRepository.existsById(id)) {
+            throw new JuiceNotFoundException("Сок с данным идендификатором не найден");
+        }
         juiceRepository.deleteById(id);
     }
 
     public void deleteAllJuices() {
-        log.info("Вся позиция соков удалена из база данных");
+        log.info("Удаление всех позиции сока из база данных");
         juiceRepository.deleteAll();
     }
 
-    public void updateJuice(String id, Juice juice) {
-        Optional<Juice> existingJuice = juiceRepository.findById(id);
+    public void updateJuice(String id, JuiceDTO juiceDTO) {
+        Optional<JuiceEntity> existingJuice = juiceRepository.findById(id);
         if (existingJuice.isPresent()) {
-            Juice updatedJuice = existingJuice.get();
-            updatedJuice.setName(juice.getName());
-            updatedJuice.setCountryOfOrigin(juice.getCountryOfOrigin());
-            updatedJuice.setColor(juice.getColor());
-            updatedJuice.setPrice(juice.getPrice());
-            updatedJuice.setVolume(juice.getVolume());
-            log.info("Сок изменился в базе данных");
+            JuiceEntity updatedJuice = juiceMapper.toEntity(juiceDTO);
+            updatedJuice.setName(updatedJuice.getName());
+            updatedJuice.setColor(updatedJuice.getColor());
+            updatedJuice.setPrice(updatedJuice.getPrice());
+            updatedJuice.setVolume(updatedJuice.getVolume());
+            log.info("Обновление сока в базе данных");
             juiceRepository.save(updatedJuice);
+        } else {
+            throw new JuiceNotFoundException("Сок с данным идендификатором не найден");
         }
     }
 
-    public Optional<Juice> findJuiceById(String id) {
-        try {
-            return juiceRepository.findById(id);
-        } catch (Exception e) {
-            throw new RuntimeException("Ошибка. Не удалось найти сок по ID " + e);
-        }
+    public JuiceDTO findJuiceById(String id) {
+        return juiceRepository.findById(id)
+                .map(juiceMapper::toDto)
+                .orElseThrow(() -> new JuiceNotFoundException("Сок с данным идендификатором не найдено"));
     }
 }
