+++ src/main/java/com/samatov/inventoryservicebar/services/VodkaService.java
@@ -1,6 +1,9 @@
 package com.samatov.inventoryservicebar.services;
 
-import com.samatov.inventoryservicebar.entities.Vodka;
+import com.samatov.inventoryservicebar.common.exceptions.notfoundexceptions.VodkaNotFoundException;
+import com.samatov.inventoryservicebar.dto.VodkaDTO;
+import com.samatov.inventoryservicebar.entities.VodkaEntity;
+import com.samatov.inventoryservicebar.mappers.VodkaMapper;
 import com.samatov.inventoryservicebar.repositories.VodkaRepository;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
@@ -10,6 +13,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -19,46 +23,55 @@ public class VodkaService {
 
     VodkaRepository vodkaRepository;
 
-    public List<Vodka> getAllVodkas() {
-        log.info("Всю позицию водки получили из база данных");
-        return vodkaRepository.findAll();
+    VodkaMapper vodkaMapper;
+
+    public List<VodkaDTO> getAllVodkas() {
+        List<VodkaEntity> vodkas = vodkaRepository.findAll();
+        log.info("Получены все позиции водки из базы данных");
+        if (vodkas.isEmpty()) {
+            throw new VodkaNotFoundException("Позиции водки не найдены в базе данных");
+        }
+        return vodkas.stream().map(vodkaMapper::toDto).collect(Collectors.toList());
     }
 
-    public void saveVodka(Vodka vodka) {
-        log.info("Водка добавлена в базу данных");
+    public void saveVodka(VodkaDTO vodkaDTO) {
+        VodkaEntity vodka = vodkaMapper.toEntity(vodkaDTO);
+        log.info("Добавление водку в базу данных");
         vodkaRepository.save(vodka);
     }
 
     public void deleteVodkaById(String id) {
-        log.info("Водка удалена из база данных");
+        log.info("Удаление водку из базы данных");
+        if (!vodkaRepository.existsById(id)) {
+            throw new VodkaNotFoundException("Водка с данным идендификатором не найдено");
+        }
         vodkaRepository.deleteById(id);
     }
 
     public void deleteAllVodkas() {
-        log.info("Вся позиция водки удалена из база данных");
+        log.info("Удаление всех позиций водки из базы данных");
         vodkaRepository.deleteAll();
     }
 
-    public void updateVodka(String id, Vodka vodka) {
-        Optional<Vodka> existingVodka = vodkaRepository.findById(id);
+    public void updateVodka(String id, VodkaDTO vodkaDTO) {
+        Optional<VodkaEntity> existingVodka = vodkaRepository.findById(id);
         if (existingVodka.isPresent()) {
-            Vodka updatedVodka = existingVodka.get();
-            updatedVodka.setName(vodka.getName());
-            updatedVodka.setCountryOfOrigin(vodka.getCountryOfOrigin());
-            updatedVodka.setColor(vodka.getColor());
-            updatedVodka.setPrice(vodka.getPrice());
-            updatedVodka.setVolume(vodka.getVolume());
-            updatedVodka.setAcoholContent(vodka.getAcoholContent());
-            log.info("Водка изменилась в базе данных");
+            VodkaEntity updatedVodka = vodkaMapper.toEntity(vodkaDTO);
+            updatedVodka.setName(updatedVodka.getName());
+            updatedVodka.setColor(updatedVodka.getColor());
+            updatedVodka.setPrice(updatedVodka.getPrice());
+            updatedVodka.setVolume(updatedVodka.getVolume());
+            updatedVodka.setAlcoholContent(updatedVodka.getAlcoholContent());
+            log.info("Обновление водки в базе данных");
             vodkaRepository.save(updatedVodka);
+        } else {
+            throw new VodkaNotFoundException("Водка с данным идендификатором не найдено");
         }
     }
 
-    public Optional<Vodka> findVodkaById(String id) {
-        try {
-            return vodkaRepository.findById(id);
-        } catch (Exception e) {
-            throw new RuntimeException("Ошибка. Не удалось найти водку по ID " + e);
-        }
+    public VodkaDTO findVodkaById(String id) {
+        return vodkaRepository.findById(id)
+                .map(vodkaMapper::toDto)
+                .orElseThrow(() -> new VodkaNotFoundException("Водка с данным идендификатором не найдено"));
     }
 }
